02_login_Signup_Flask/requirements.txt
Flask==2.2.2
PyJWT==2.8.0
PyJWT==2.8.0
pymongo==4.2.0
python_bcrypt==0.3.2
